+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 
 ## Release 3.7
 
+* #449 Excludes in overlays: support removing attributes and relationships via overlays
 * #450 Minimize Cayenne footprint - remove references to "Property" from API signatures
 * #451 Remove API deprecated in 3.4 or earlier
 * #452 Replace SelectBuilder.property(..) with overlay-based API